Fix: Loss weights were not passed to the loss function.
Minors:
- Auto-casting and float32 mm precision trade-offs are now applied only to the backbone. Other critical parts will remain as true float32.
- Fix the validation iou/giou issue. Now those ious are computed only for within a GT interval.

+++ ltvu/lit/data.py
@@ -69,7 +69,8 @@ class LitVQ2DDataModule(L.LightningDataModule):
             flat_anns = generate_flat_annotations_vq2d(all_anns)
             assert len(flat_anns) == desired_num_anns, f'Split {split} has {len(flat_anns)} annotations, expected {desired_num_anns}'
             p_ann = self.p_anns_dir / f'vq_v2_{split}_anno.json'
-            json.dump(flat_anns, p_ann.open('w'))
+            if not p_ann.exists():
+                json.dump(flat_anns, p_ann.open('w'))
             for ann in flat_anns:
                 video_uids.add(ann['video_uid'])
                 clip_uids.add(ann['clip_uid'])
@@ -120,7 +121,6 @@ class LitVQ2DDataModule(L.LightningDataModule):
 
         bbox_scale = torch.tensor([h, w, h, w], device=device)[None, None]  # [1,1,4]
         bboxes_px = gt_bboxes * bbox_scale  # [b,t,4], pixel space
-        # bboxes_px = bboxes_px[..., [1, 0, 3, 2]]  # [b,t,4], yxyx format
         seg_queue, bboxes_queue = [], []
         # DON'T FLATTEN AND PRALLELIZE THIS LOOP
         #    AS t IS TREATED AS BATCH AXIS
@@ -132,17 +132,6 @@ class LitVQ2DDataModule(L.LightningDataModule):
             # [t,c,h,w] -> [c,h,w], [b,t,4] -> [t,4] -> [t,N=1,4]
             segment_aug, bboxes_aug_px = self.transform_clip(segment[i], bboxes_px[i, :, None])
             bboxes_aug_px = bboxes_aug_px.squeeze(1)  # [t,4]
-            # bboxes_aug = bboxes_aug[..., [1, 0, 3, 2]]  # [t,4], xyxy format
-            # imgs = []
-            # for t in range(segment_aug.shape[0]):
-            #     img = Image.fromarray(segment_aug[t].permute(1, 2, 0).mul(255).byte().cpu().numpy())
-            #     draw = ImageDraw.Draw(img)
-            #     bbox = bboxes_aug[t].mul(torch.tensor([w, h, w, h], device=device))
-            #     draw.rectangle(bbox.tolist(), outline='red', width=5)
-            #     imgs.append(img)
-            # p_img = Path(f'outputs/dataset/seg-{device}-{i}.gif')
-            # imgs[0].save(p_img, save_all=True, append_images=imgs[1:], duration=100, loop=0)
-            # print(f'bbox: {p_img} {bbox}')
             seg_queue.append(segment_aug)
             bboxes_queue.append(bboxes_aug_px)
         segment = torch.stack(seg_queue, dim=0)
